Rename external matchers package to 'extensions'
